+++ src/compas_slicer/slicers/interpolation_slicer.py
@@ -46,8 +46,7 @@ class InterpolationSlicer(BaseSlicer):
         params_list = get_interpolation_parameters_list(n)
         logger.info('%d paths will be generated' % n)
 
-        max_dist = get_param(self.parameters, key='vertical_layers_max_centroid_dist', defaults_type='layers')
-        vertical_layers_manager = VerticalLayersManager(max_dist, avg_layer_height)
+        vertical_layers_manager = VerticalLayersManager(avg_layer_height)
 
         # create paths + layers
         with progressbar.ProgressBar(max_value=len(params_list)) as bar: